+++ packages/utils/src/index.ts
@@ -11,3 +11,4 @@ export * from "./math";
 export * from "./signatures";
 export * from "./basic";
 export * from "./hexStrings";
+export * from "./transactionManager";
