+++ packages/utils/src/transactionManager.ts
@@ -1,0 +1,17 @@
+export type InvariantTransactionData = {
+  user: string;
+  router: string;
+  sendingAssetId: string;
+  receivingAssetId: string;
+  receivingAddress: string;
+  sendingChainId: number;
+  receivingChainId: number;
+  callData: string;
+  transactionId: string;
+};
+
+export type TransactionData = InvariantTransactionData & {
+  amount: string;
+  expiry: string;
+  blockNumber: number;
+};
